+++ core/dbt/parser/schema_test_builders.py
@@ -41,16 +41,21 @@ def get_nice_schema_test_name(
     clean_flat_args = [re.sub('[^0-9a-zA-Z_]+', '_', arg) for arg in flat_args]
     unique = "__".join(clean_flat_args)
 
-    cutoff = 32
-    if len(unique) <= cutoff:
-        label = unique
+    # for the file path + alias, the name must be <64 characters
+    # if the full name is too long, include the first 30 identifying chars plus
+    # a 32-character hash of the full contents
+
+    test_identifier = '{}_{}'.format(test_type, test_name)
+    full_name = '{}_{}'.format(test_identifier, unique)
+
+    if len(full_name) >= 64:
+        test_trunc_identifier = test_identifier[:30]
+        label = hashlib.md5(full_name.encode('utf-8')).hexdigest()
+        short_name = '{}_{}'.format(test_trunc_identifier, label)
     else:
-        label = hashlib.md5(unique.encode('utf-8')).hexdigest()
+        short_name = full_name
 
-    filename = '{}_{}_{}'.format(test_type, test_name, label)
-    name = '{}_{}_{}'.format(test_type, test_name, unique)
-
-    return filename, name
+    return short_name, full_name
 
 
 @dataclass
@@ -185,7 +190,7 @@ class TestBuilder(Generic[Testable]):
         r'(?P<test_name>([a-zA-Z_][0-9a-zA-Z_]*))'
     )
     # kwargs representing test configs
-    MODIFIER_ARGS = ('severity', 'tags', 'enabled')
+    MODIFIER_ARGS = ('severity', 'tags', 'enabled', 'store_failures')
 
     def __init__(
         self,
@@ -231,6 +236,10 @@ class TestBuilder(Generic[Testable]):
         self.compiled_name: str = compiled_name
         self.fqn_name: str = fqn_name
 
+        # use hashed name as alias if too long
+        if compiled_name != fqn_name:
+            self.modifiers['alias'] = compiled_name
+
     def _bad_type(self) -> TypeError:
         return TypeError('invalid target type "{}"'.format(type(self.target)))
 
@@ -271,12 +280,18 @@ class TestBuilder(Generic[Testable]):
     def enabled(self) -> Optional[bool]:
         return self.modifiers.get('enabled')
 
+    def alias(self) -> Optional[str]:
+        return self.modifiers.get('alias')
+
     def severity(self) -> Optional[str]:
         sev = self.modifiers.get('severity')
         if sev:
             return sev.upper()
         else:
             return None
+
+    def store_failures(self) -> Optional[bool]:
+        return self.modifiers.get('store_failures')
 
     def tags(self) -> List[str]:
         tags = self.modifiers.get('tags', [])