+++ test/integration/029_docs_generate_tests/test_docs_generate.py
@@ -986,9 +986,10 @@ class TestDocsGenerate(DBTIntegrationTest):
             'vars': {},
             'tags': [],
             'severity': 'ERROR',
+            'store_failures': None,
             'full_refresh': None,
             'database': None,
-            'schema': None,
+            'schema': 'dbt_test__audit',
             'alias': None,
         }
         result.update(updates)
@@ -1054,6 +1055,7 @@ class TestDocsGenerate(DBTIntegrationTest):
         snapshot_path = self.dir(os.path.join('snapshot', 'snapshot_seed.sql'))
 
         my_schema_name = self.unique_schema()
+        test_audit_schema = my_schema_name + '_dbt_test__audit'
 
         if model_database is None:
             model_database = self.alternative_database
@@ -1350,7 +1352,7 @@ class TestDocsGenerate(DBTIntegrationTest):
                     'relation_name': None,
                     'resource_type': 'test',
                     'root_path': self.test_root_realpath,
-                    'schema': my_schema_name,
+                    'schema': test_audit_schema,
                     'database': self.default_database,
                     'tags': ['schema'],
                     'meta': {},
@@ -1439,7 +1441,7 @@ class TestDocsGenerate(DBTIntegrationTest):
                     'relation_name': None,
                     'resource_type': 'test',
                     'root_path': self.test_root_realpath,
-                    'schema': my_schema_name,
+                    'schema': test_audit_schema,
                     'database': self.default_database,
                     'tags': ['schema'],
                     'meta': {},
@@ -1484,7 +1486,7 @@ class TestDocsGenerate(DBTIntegrationTest):
                     'relation_name': None,
                     'resource_type': 'test',
                     'root_path': self.test_root_realpath,
-                    'schema': my_schema_name,
+                    'schema': test_audit_schema,
                     'database': self.default_database,
                     'tags': ['schema'],
                     'meta': {},